refactor: refactor routes
