+++ core/routes/404.js
@@ -1,0 +1,11 @@
+const errorHandler = (error, _, response, next) => {
+  console.log("hre", error);
+  response.render("pages/404", {
+    title: "OOPS! \n" + "404 PAGE NOT FOUND",
+    message:
+      "The page you’re looking for does not exist or has been removed.\n" +
+      'You can proceed to our <a href="/">home page</a>.',
+  });
+};
+
+module.exports = errorHandler;
