+++ core/routes/home.js
@@ -1,0 +1,9 @@
+const router = require("express").Router();
+
+router.get("/", (_, response) => {
+  response.render("pages/home", {
+    title: "Home",
+  });
+});
+
+module.exports = router;
